Reorder refinery directory (#2)
This PR rearranges the code in this gem slightly, moving "library code" over into `lib`, and keeping the system dependency booting code in `system`:

I've found this way of arranging code to be nicer to work with in https://github.com/radar/bix. 

I've tested that this PR works by using the included `test.rb`script.